pyproject.toml
[project]
name = "ring-textilservice-prototype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "awscli>=1.42.16",
    "awsglue-dev>=2021.12.30",
    "boto3>=1.40.16",
    "pandas[excel]>=2.3.2",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "pypandoc==1.5",
    "pyspark>=4.0.0",
]
